server/requirements.txt
aiosql==3.2.0
aiosqlite==0.16.0
alembic==1.4.3
appdirs==1.4.4
asyncpg==0.21.0
black==20.8b1
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
databases==0.4.1
ecdsa==0.14.1
fastapi==0.62.0
flake8==3.8.4
h11==0.11.0
isort==5.6.4
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
PyPika==0.46.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
regex==2020.11.13
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.13.1
